Move scraping code into a separate file.
All the scraping functions were included in the users_controller.rb file
and there was a good deal of code that could be extracted from the
various scraping methods.  This commit refactored the logic for calling
the appropriate scraper by creating a base Scraper class along with
various subclasses that implement the 'doScrape' method to perform
site-specific scraping.

I've manually tested that all the functions still work, but this
refactoring might also make it easier to create some automated tests.