Changes to site layout in progress.
I've added a new Search controller that has a simpler UI, but
it isn't hooked up to the Eventful API yet.  I'm experimenting
with changes to the CSS stylesheets and the various layouts, but
don't think I've improved anything yet.  I'd like to have a
search form available at the top of every page with results
displayed below.