Add validation and test for User model.
Added validation to require EITHER the date
OR the city field to be filled in before performing a search.
I think this validation belongs in the model rather than the
controller where we had been doing a hack of inserting 'usa' if the city field
was blank.

All the tests pass now, but we need to refactor so that we don't
actually ping spotify during our testing.